Refactor mock servlet API classes to mock package
